seperated out into helper functions.
